feat: update tRPC configuration
