--- apps/web/src/app/providers/TrpcProvider.tsx
@@ -1,30 +1,0 @@
-'use client';
-
-import { getQueryClient, trpc } from '@/fsd/shared';
-
-import { QueryClientProvider } from '@tanstack/react-query';
-import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
-import { httpBatchLink } from '@trpc/client';
-import { useState } from 'react';
-
-export function TrpcProvider({ children }: { children: React.ReactNode }) {
-	const queryClient = getQueryClient();
-	const [trpcClient] = useState(() =>
-		trpc.createClient({
-			links: [
-				httpBatchLink({
-					url: '/api/trpc',
-				}),
-			],
-		}),
-	);
-
-	return (
-		<trpc.Provider client={trpcClient} queryClient={queryClient}>
-			<QueryClientProvider client={queryClient}>
-				{children}
-				<ReactQueryDevtools initialIsOpen={false} />
-			</QueryClientProvider>
-		</trpc.Provider>
-	);
-}
